requirements.txt
youtube-transcript-api
anthropic
python-dotenv
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.2.0
